requirements.txt
streamlit
openai
requests
python-dotenv
streamlit-lottie
pypdf
pytesseract
pdf2image
reportlab
matplotlib
pandas
pillow